mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP Server for Planner AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
    "fastapi>=0.109.0",
    "fastmcp-http",
]